requirements.txt
kagglehub==0.1.7
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv
langchain-openai==0.2.0
langchain-chroma==0.2.0
transformers==4.30.2
gradio==3.39.0
notebook==5.7.8
ipywedges==0.5.0